requirements.txt
langchain
langchain_openai
langchain-core
langchain_community
python-dotenv
langchain-google-genai
langchain-chroma
langchain-text-splitters

streamlit